refactor: group into TaxProvidersAccordion
